:checkmark: move test_FBS_methods to its own build and skip some duplicative methods
